+++ esperanto-tools-libs/src/KernelLaunch.cpp
@@ -155,7 +155,7 @@ EventId RuntimeImp::doKernelLaunch(StreamId streamId, KernelId kernelId, const s
                << cmdPtr->pointer_to_args << ", PC: 0x" << cmdPtr->code_start_address << ", shireMask: 0x"
                << options.shireMask_;
   auto& commandSender = find(commandSenders_, getCommandSenderIdx(streamInfo.device_, streamInfo.vq_))->second;
-  commandSender.send(Command{cmdBase, commandSender, event, event, false, true});
+  commandSender.send(Command{cmdBase, commandSender, event, event, streamId, false, true});
 
   Sync(event);
   return event;