+++ esperanto-tools-libs/src/ProfilerImp.h
@@ -36,86 +36,23 @@ public:
   void record(const ProfileEvent&) override {
     // Intentionally unimplemented
   }
+  bool isDummy() const override {
+    return true;
+  }
 };
 
-// Regluar implementation
+// Regular implementation
 class ProfilerImp : public IProfilerRecorder {
 public:
   // IProfiler interface
-  void start(std::ostream& outputStream, OutputType outputType) override {
-    if (recording_) {
-      throw Exception("Profiler was already started");
-    }
-    recording_ = true;
-    ioThread_ = std::thread(std::bind(&ProfilerImp::ioThread, this, outputType, &outputStream));
-    ProfileEvent evt(Type::Instant, Class::StartProfiling);
-    evt.setExtras({{"version", kCurrentVersion}});
-    record(evt);
-  }
-
-  void stop() override {
-    if (recording_) {
-      ProfileEvent evt(Type::Instant, Class::EndProfiling);
-      record(evt);
-      recording_ = false;
-      cv_.notify_one();
-      ioThread_.join();
-    }
-  }
-
-  void record(const ProfileEvent& event) override {
-    if (!recording_)
-      return;
-    SpinLock lock{mutex_};
-    auto wasEmpty = events_.empty();
-    events_.push(event);
-    lock.unlock();
-    if (wasEmpty) {
-      cv_.notify_one();
-    }
-  }
-
-  ~ProfilerImp() override {
-    if (recording_) {
-      stop();
-    }
-  }
-
-  void ioThread(OutputType outputType, std::ostream* stream) {
-    std::variant<std::monostate, cereal::JSONOutputArchive, cereal::PortableBinaryOutputArchive> archive;
-    switch (outputType) {
-    case OutputType::Json:
-      archive.emplace<cereal::JSONOutputArchive>(*stream);
-      break;
-    case OutputType::Binary:
-      archive.emplace<cereal::PortableBinaryOutputArchive>(*stream);
-      break;
-    default:
-      throw Exception("Unknown profiler output type");
-    }
-    SpinLock lock{mutex_};
-    while (recording_ || !events_.empty()) {
-      if (events_.empty()) {
-        cv_.wait(lock, [this] { return !events_.empty() || !recording_; });
-      }
-      if (!events_.empty()) {
-        auto evt = events_.front();
-        events_.pop();
-        lock.unlock();
-        std::visit(
-          [&evt](auto&& arch) {
-            using T = std::decay_t<decltype(arch)>;
-            if constexpr (!std::is_same_v<T, std::monostate>) {
-              arch(evt);
-            }
-          },
-          archive);
-        lock.lock();
-      }
-    }
-  }
+  void start(std::ostream& outputStream, OutputType outputType) override;
+  void stop() override;
+  void record(const ProfileEvent& event) override;
+  ~ProfilerImp() override;
 
 private:
+  void ioThread(OutputType outputType, std::ostream* stream);
+
   std::mutex mutex_;
   std::queue<ProfileEvent> events_;
   std::condition_variable cv_;