+++ esperanto-tools-libs/src/dma/MemcpyListD2HAction.cpp
@@ -71,8 +71,8 @@ bool MemcpyListD2HAction::update() {
   RT_VLOG(MID) << ">>> Alloc cmaPtr: " << std::hex << cmaPtr << " associated events: " << stringizeEvents(syncEvents);
 
   // set the proper data once the builder has been filled
-  ctx_.commandSender_.sendBefore(ctx_.eventId_,
-                                 {builder.build(), ctx_.commandSender_, cmdEvt, ctx_.eventId_, true, true});
+  ctx_.commandSender_.sendBefore(
+    ctx_.eventId_, {builder.build(), ctx_.commandSender_, cmdEvt, ctx_.eventId_, ctx_.stream_, true, true});
 
   // release the buffer once the command has been completed
   ctx_.eventManager_.addOnDispatchCallback(