+++ esperanto-tools-libs/src/server/Server.h
@@ -22,7 +22,7 @@ public:
   explicit Server(const std::string& socketPath, std::unique_ptr<dev::IDeviceLayer> deviceLayer, Options options);
   ~Server();
   void removeWorker(Worker* worker);
-  void setProfiler(std::unique_ptr<profiling::IProfilerRecorder> profiler) {
+  void setProfiler(std::unique_ptr<profiling::IProfilerRecorder>&& profiler) {
     runtime_->setProfiler(std::move(profiler));
   }
   profiling::IProfilerRecorder* getProfiler() const {