+++ esperanto-tools-libs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [[_TOC_]]
 ## [unreleased]
 ### Added
+- [SW-19991] Forward server tracing events to clients if tracing is enabled
 ### Changed
 ### Deprecated
 ### Removed